# Eliminate FFmpeg Dependency - Pure Rust WebRTC Implementation